pyproject.toml
[project]
name = "kantan-agents"
version = "0.1.0"
description = "kantan-agents is a thin, opinionated wrapper around the OpenAI Agents SDK that makes observability and evaluation \u201cjust happen\u201d by default."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kantan-llm>=0.1.6",
    "openai-agents>=0.6.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
